+++ src/main/java/dev/astranfalio/teioc/repository/InternRepository.java
@@ -4,6 +4,6 @@ import dev.astranfalio.teioc.entity.InternEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface InternRepository extends JpaRepository<InternEntity, Long> {
+public interface InternRepository extends JpaRepository<InternEntity, Integer> {
 
 }
